+++ src/activeNetworkInterfacesModel.cpp
@@ -70,8 +70,8 @@ private:
 	ActiveNetworkInterfacesModel* const q_ptr{ nullptr };
 	Q_DECLARE_PUBLIC(ActiveNetworkInterfacesModel);
 
-	hive::widgetModelsLibrary::NetworkInterfacesListModel _model{};
-	std::unordered_set<la::networkInterface::Interface::Type> _allowedInterfaceTypes{};
+	hive::widgetModelsLibrary::NetworkInterfacesListModel _model{ true };
+	std::unordered_set<la::networkInterface::Interface::Type> _allowedInterfaceTypes{ la::networkInterface::Interface::Type::Loopback };
 };
 
 ActiveNetworkInterfacesModel::ActiveNetworkInterfacesModel(QObject* parent)